backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic[email]>=2.0.0
qrcode[pil]>=7.4.0
reportlab>=4.0.0
PyJWT>=2.8.0
pytest>=8.2.0
